Explicit cast in make_pair
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@47572 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/test/test_ncbi_ftp_download.cpp
@@ -250,7 +250,7 @@ size_t CListProcessor::Run(void)
         if (!line.empty()) {
             cerr.flush();
             cout << left << setw(60) << line << endl;
-            CTar::TFile file = make_pair(line, 0);
+            CTar::TFile file = make_pair(line, (Uint8) 0);
             m_Dlcbdata->Append(&file);
             n++;
         }
